added if statement for dir checking

+++ setup.sh
@@ -10,7 +10,9 @@ if [[ ":$PATH:" != *":$HOME/bin:"* ]]; then
 	export PATH="$PATH:$HOME/bin"
 fi
 
-ln -sf "$PWD/scripts" "$HOME/scripts"
+if [[ ! -L "$HOME/scripts" ]]; then
+    ln -sf "$PWD/scripts" "$HOME/scripts"
+fi
 
 for i in fish newsboat tmux vim ; do
 	cd $i && ./setup.sh
